restructure in prep for graphml exporter
