+++ formats/src/test/scala/overflowdb/formats/graphml/GraphMLTests.scala
@@ -1,4 +1,4 @@
-package overflowdb.formats
+package overflowdb.formats.graphml
 
 import org.scalatest.matchers.should.Matchers._
 import org.scalatest.wordspec.AnyWordSpec
